requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
bleach==1.5.0
bokeh==0.12.5
chest==0.2.3
cloudpickle==0.2.2
configparser==3.5.0
cycler==0.10.0
dask==0.14.1
decorator==4.0.11
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
futures==3.0.5
h5py==2.6.0
HeapDict==1.0.0
html5lib==0.999
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.9.5
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==1.2.2
locket==0.2.0
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.7.3
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
notebook==4.3.1
numexpr==2.6.2
numpy==1.12.0
olefile==0.44
pandas==0.19.2
pandocfilters==1.4.1
partd==0.3.7
pathlib2==2.2.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycairo==1.10.0
Pygments==2.2.0
pyparsing==2.1.4
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.13.0
root-numpy==4.4.0.dev0
rootpy==0.8.1.dev0
scandir==1.5
scikit-image==0.12.3
scikit-learn==0.18.1
scikit-optimize==0.3
scipy==0.18.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
stevedore==1.18.0
subprocess32==3.2.7
tables==3.3.0
terminado==0.6
testpath==0.3
Theano==0.8.2
toolz==0.8.2
tornado==4.4.2
tqdm==4.11.2
traitlets==4.3.2
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wcwidth==0.1.7
widgetsnbextension==1.2.6
